Moved certain parts of the GUI script to their own files
